Fix incorrect index type in vertex data struct resulting mesh artefacts

+++ projects/app/src/rage/grm/geometry.h
@@ -21,7 +21,7 @@ namespace rage
 		pVoid				Vertices;
 		grcIndices_t		Indices;
 		u32					VertexCount;
-		u16					IndexCount;
+		u32					IndexCount;
 		grcVertexFormatInfo	Info;
 	};
